requirements.txt
ipdb==0.11
Keras==2.1.5
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
matplotlib==2.2.3
numpy==1.14.5
Pillow==5.2.0
pip==18.0
tensorboard==1.6.0
tensorflow-gpu
tensorflow-tensorboard==0.4.0
scikit-learn==0.19.1
opencv-python
sklearn
scipy


